requirements.txt
openai>=1.0.0
asyncpg>=0.29.0
pytest>=7.0.0
python-dotenv>=1.0.0
